Quite a few bugfixes, statically linked executables get further (#15)

+++ src/felix86/hle/syscall.cpp
@@ -177,7 +177,7 @@ void felix86_syscall(Emulator* emulator, ThreadState* state) {
     }
     }
 
-    if (result == -1) {
+    if ((i64)result == -1) {
         result = -errno;
     }
